fix cucumber hooks
