+++ src/test/java/nz/ac/auckland/se754/web/stepdefs/ChromeWebDriverManager.java
@@ -1,0 +1,66 @@
+package nz.ac.auckland.se754.web.stepdefs;
+
+import io.cucumber.java.After;
+import io.cucumber.java.AfterStep;
+import io.cucumber.java.Before;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+
+import java.util.concurrent.TimeUnit;
+
+public class ChromeWebDriverManager {
+    private static WebDriver driver;
+
+    public static WebDriver getDriver() {
+        return driver;
+    }
+
+    @Before(order = 0)
+    public void setup() {
+        ChromeOptions options = new ChromeOptions();
+        String osName = System.getProperty("os.name").toLowerCase();
+        String osArch = System.getProperty("os.arch").toLowerCase();
+
+        if (System.getenv().getOrDefault("headless", "false").equals("true")) {
+            options.addArguments("--headless");
+            System.setProperty("webdriver.chrome.driver", "webdrivers/linux/chromedriver");
+        } else if (osName.contains("mac")) {
+            if (osArch.equals("aarch64")) {
+                System.setProperty("webdriver.chrome.driver", "webdrivers/macos/aarch64/chromedriver");
+            } else {
+                System.setProperty("webdriver.chrome.driver", "webdrivers/macos/x64/chromedriver");
+            }
+        } else if (osName.contains("win")) {
+            System.setProperty("webdriver.chrome.driver", "webdrivers/win/chromedriver.exe");
+        }
+
+        options.addArguments("--no-sandbox");
+        options.addArguments("--disable-dev-shm-usage");
+        options.addArguments("--disable-gpu");
+        options.addArguments("--window-size=1920,1200");
+        options.addArguments("--ignore-certificate-errors");
+        options.addArguments("--disable-extensions");
+
+        driver = new ChromeDriver(options);
+        driver.manage().window().maximize();
+        driver.manage().timeouts().implicitlyWait(45, TimeUnit.SECONDS);
+    }
+
+    @AfterStep
+    public void afterEachStep() {
+        // to make the test at human speed
+        if (System.getenv().getOrDefault("headless", "false").equals("false")) {
+            try {
+                Thread.sleep(2000);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+        }
+    }
+
+    @After
+    public void tearDown() {
+        driver.quit();
+    }
+}
