+++ src/test/java/nz/ac/auckland/se754/web/stepdefs/VocabDefinitionsStepDefinitions.java
@@ -1,18 +1,13 @@
 package nz.ac.auckland.se754.web.stepdefs;
 
-import io.cucumber.java.After;
-import io.cucumber.java.AfterStep;
 import io.cucumber.java.Before;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import nz.ac.auckland.se754.web.pages.LessonPage;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
 
 import java.util.ArrayList;
-import java.util.concurrent.TimeUnit;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -21,49 +16,10 @@ public class VocabDefinitionsStepDefinitions {
     private LessonPage lessonPage;
     private String word;
 
-
-    @Before
-    public void setup() {
-        ChromeOptions options = new ChromeOptions();
-        String osName = System.getProperty("os.name").toLowerCase();
-
-        if (System.getenv().getOrDefault("headless", "false").equals("true")) {
-            options.addArguments("--headless");
-            System.setProperty("webdriver.chrome.driver", "webdrivers/linux/chromedriver");
-        } else if(osName.contains("mac")) {
-            System.setProperty("webdriver.chrome.driver", "webdrivers/macos/chromedriver");
-        } else if(osName.contains("win")) {
-            System.setProperty("webdriver.chrome.driver", "webdrivers/win/chromedriver.exe");
-        }
-
-        options.addArguments("--no-sandbox");
-        options.addArguments("--disable-dev-shm-usage");
-        options.addArguments("--disable-gpu");
-        options.addArguments("--window-size=1920,1200");
-        options.addArguments("--ignore-certificate-errors");
-        options.addArguments("--disable-extensions");
-
-        driver = new ChromeDriver(options);
-        driver.manage().window().maximize();
-        driver.manage().timeouts().implicitlyWait(45, TimeUnit.SECONDS);
+    @Before("@vocabDefinitions")
+    public void before() {
+        driver = ChromeWebDriverManager.getDriver();
         lessonPage = new LessonPage(driver);
-    }
-
-    @AfterStep
-    public void afterEachStep() {
-        // to make the test at human speed
-        if (System.getenv().getOrDefault("headless", "false").equals("false")) {
-            try {
-                Thread.sleep(2000);
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
-        }
-    }
-
-    @After
-    public void tearDown() {
-        driver.quit();
     }
 
     @Given("I am in a lesson")
@@ -133,7 +89,6 @@ public class VocabDefinitionsStepDefinitions {
         String definition = lessonPage.getDefinition();
         String expectedDefinition = "valid definition";
         assertEquals(definition, expectedDefinition);
-
     }
 
     @Then("I should see an alert message pop-up saying unable to find definition")
@@ -148,7 +103,6 @@ public class VocabDefinitionsStepDefinitions {
         String example = lessonPage.getExample();
         String expectedExpected = "valid example";
         assertEquals(example, expectedExpected);
-
     }
 
     @Then("I should see an alert message pop-up saying unable to find any examples")
