+++ src/test/resources/features/login.feature
@@ -1,3 +1,4 @@
+@login
 Feature: login
 
   Scenario: login
